Resolved error on free after allocating wrong size
size was being used incorrectly and was used for multiple purposes
after the request code was rewritten. size is now only used to store
the length of the request buffer. idx now accomplishes the task of
tracking where in the buffer we are and, thus, where to store the
next set of values.