Fatal error when destination refuses connection
We need to be able to handle when we successfully establish a
connection with the SOCKS server but the destination server
fails to accept the connect from the SOCKS server. In this case,
the SOCKS server returns a bind address and port consisting of
null bytes. We now check for this and inform the user.