Added a solution to be called if peer disconnects
When the client becomes disconnected from the peer, we want a way
to run an arbitrary "solution" that will help to reestablish the
connection. By providing a command line option, the client will
run this provided solution and attempt to reconnect to the peer.

This implementation is based on the assumption that the client may
lose connection to the internet for a known reason and by running an
executable/script/etc the connection may be reestablished and the peer
may be contacted again and the request-response conversation may
continue.